Implement ORCiD login auto-redirect for OSF sign-up
It is very hard use the client set that is stored in the flow
context simply to find the one URL for ORCiD login. Thus, the
pre-login check action loads the full set from the flow, finds
the URL and then put theminto the dedicated login context.

Created a layout-less template page just for the redirection
using the meta tag with delay set to 0.

If the URL is not ready or the service is not eligible, log
the error and redirect flow to the main login page.