Implement OSF URL prop and put it in the flow context
CASv4.1.x-based oldCAS uses JSP as the template engine with Spring.
Settings defined in .properties files can be accessed directly in
JSPs. However, it is no longer the case for CASv6.2.x-based newCAS
that uses Thymeleaf as the template engine (with Spring of course).
Thus, server-specific URLs (i.e. production, test and stage) need
to be manually put into the flow context for templates to access.