Fix an issue where pac4j-cas instn login URL were not put in context
This is a similar solution to the ORCiD login URL except there are
multiple cas clients in this case. Thus, the institution pre-login
check action has been expanded to init w/ pre-defined cas clients,
retrieve respective URLs from the flow and re-put them in as a map
(which was a set) for easier template access.

In addition, fixed the script on the institution sso view template
by using <script th:inline="javascript"> so that thymeleaf can now
parse and use flow context vars correctly.