Adjust for Tomcat's lower-case HTTP headers
Apache Tomcat has an unwanted but have-to-live-with behavior that
it changes all HTTP headers to lower-case. Although, HTTP headers
are not case-sensitive according to specifications, Tomcat should
have made headers intact. This limitation has a side-effect that
our attributes transforming and parsing process has to be updated
for Shib-SAML based institution SSO to work.