Implement pre-flow-init action to prepare server-specific settings
OSF server-specific settings such as OSF URLs need to be configured
at the very beginning of the login flow so that ThymeLeaf templates
can be rendered correctly across the flow state machine.

Previously, it was configured at the start of the non-interactive
action where any rendering attempt before it would fail with empty
pages rendered. This is fixed now by overriding the login flow init
process. A pre-init action with what we did in the non-interactive
action before is injected onto the top of the start action list.