Rewrite 400s templates with the flowless layout and header
HTTP 400 errors are not expected during normal authentication and
authorization flow between OSF CAS and OSF. Sometimes they do not
trigger login or logout flow init and are handled and routed to
built-in templates directly by Tomcat. This unfortunately breaks
OSF CAS customized fragment templates that relies on web flow
init. The fix is to use a flowless layout and header, then replace
the no-service "/login?" href with one with "errorSource=4XX" as
query param so that the pre-login check action can handle.