Rewrite revoke endpoint (partially) to support OSF (the oldCAS way)
Vanilla newCAS revoke endpoint requires the token ID, the client
ID and  an authenticated session for the request to be validated
and successful. In addition, it only supports revoking one token
(or multiple if refresh token). This doesn't fit current OSF.

Added new / rewrote existing validators to support two types of
revokation used by OSF: one token (or multiple if refresh token)
and all tokens of a client. Updated configurations and rewrote
the controller and related utility methods.

Although there are lots of OSF customizations for this endpoint,
the backend for database IO (i.e. reading from and writing to the
ticket registry) has not been changed. This is a key different
from oldCAS where we wrote everything by ourselves, which led to
a bug where ticket deletetions sometimes fail due to DB error.

One extra note: "partially" means the second feature (removing all
tokens of a given client) is not implemented. For now, it returns
501 HTTP NOT IMPLEMENTED.