Enable online or offline via access type
Vanilla newCAS generates refresh token only if the service allows
it, which does not support the oldCAS feature where the choice can
be made with request params. In addition, oldCAS has a complex way
of defining the type of the code and token that needs to be trimmed
down.

Thus, oldCAS "TokenType" is re-designed into "OsfOAuth20CodeType"
in newCAS, with four values: OFFLINE, ONLINE, PERSONAL and VANILLA.
The first two are used by authorization code, the third one is used
only by access tokens generated from OSF PAT, and the last one is
for all OAuth generated access and refresh tokens.

All related classes are updated so that newCAS now supports the
"access_type" request param for the authorization endpoint. The
default value is "online", which only allows the code to generate
one access token; the "offline" value must be set explicitly and
it allows the code to generate a refresh token in addition to the
access token.