/profile endpoint now releases scopes
Vanilla newCAS does not release scopes which OSF API needs to check
against. Thus, scope set is retrieved from the access token and added
to the profile. In addition, a workaround has been implemented to fix
a bug in the ObjectMapper.writeValueAsString() where one element set
or list is incorrectly converted to a plain string in the response.
Finally, more attributes including the client id, the service id and
authentication attributes are returned.