Implement scope verification with new view model context
Vanilla newCAS OAuth 2.0 module does not verify scopes. Thus, added
a check that uses JPA OSF DAO to verify that the scopes provided in
the /authorize request are valid OSF scopes.

The vanilla error view does not have a model that stores helpful
error information for users when the /authorize request fails. In
addition, it also lacks the context of OSF URLs that are server env
dependent. Thus, created a new class that stores all these data and
is sent to the front-end template.

With the model context issue fixed, the error page (currently only
used by invalid scope due to validator design) are re-written to
share the newCAS OSF-like design.