Implement OsfPostgresServiceRegistry to support OSF developer apps
OsfPostgresServiceRegistry is a customized service registry that
uses OSF JPA data access model to load OSF dev apps into CAS and
transform them into the built-in OAuth 2.0 registered services.

In addition, this new service registry is injected into the CAS
implementation so that the built-in service manager can manage
and access them as if they were built-in ones. Thus, any class
that has access to the service manager can access OSF dev apps.

`build.gradle` and `spring/factories` are updated as well.