Use default layout to handle "/error" (i.e. server error page)
"/error" is the end view state whenever an unexpected and unhandled
exception occurs, which can be triggered in both web flows that are
aware or unaware of the OSF. Thus, the template must use the default
layout and header to avoid rendering issues when the OSF is not in
context (though the chance is quite low).

In addition, URLs on this page have been updated to redirect users
back to OSF-aware pages so that they can navigate back to the OSF or
login again with the OSF as the service.