Use absolute URL for logout service URLs
Unlike other staging servers (including the test), staging 1 has a
bug where CAS redirects the logout flow to 127.0.0.1 if a relative
URL is used as the service URL. Although this is a DevOps issue,
the current fix is to use absolute URLs for all logout service URLs
in all templates.

In order for the ThymeLeaf templates to access "cas.server.name" in
"cas.properties", a bean must be registered at the Spring Application
Context. The current fix only give templates the full CAS login URL.

This solution has also made it possible to solve the server-specific
URLs issue properly rather than splitting OSF-aware/unaware flows,
which will be a future improvement.