Fix bean creation failure
"scopedTarget.oauthHandlerInterceptorAdapter" sometimes has a hard
time being initialized, throwing "BeanCreationNotAllowedException".

The error is "Singleton bean creation not allowed while singletons
of this factory are in destruction". This only happens on stage 2,
stage 3 and test, and only when all three are being deployed at the
same time.

We were not able to figure out why with 100% certainty. The guess
is that the annotation "@RefreshScope" may have caused the issue.